requirements.txt
altgraph==0.17.3
EasyProcess==1.1
entrypoint2==1.1
jeepney==0.8.0
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
PyAutoGUI==0.9.54
pygame==2.5.1
PyGetWindow==0.0.9
pyinstaller==5.13.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python3-xlib==0.15
pytweening==1.0.7
